frontend/requirements.txt
numpy
opencv-python
pillow
python-multipart
uvicorn
streamlit
fastapi
OmegaConf